fix(Dialog): Removed extra props
